Fix broken Names to generated LLVM functions
Function expressions and thunks generate functions, and use names
generated from IRBuilderT. The first problem is that the generated
functions are `lift`ed out of IRBuilderT, so each reset the name supply.
Given duplicate names are super possible, it follows that the same Name
in a straight ModuleBuilder vs an IRBuilderT setting can serialize
differently.

To fix the broken names, carry around our own State, but this time
innermost in the stack. Then it is preserved across `lift`s. The names
generated have proper string names.

Also in this commit is proper usage of getelementptr. Before I was
providing the type as "ptr", thinking it was asking for the input type,
but it is instead asking for the pointee type. Now that it is given
proper pointee types, I can also give it two indices.